Internal BTree implementation to replace the current treemap/treeset
This implementation was inspired by
"github.com/tonsky/persistent-sorted-set" for clojure and adapted so
that it is as efficient as possible in Go.

Additional modifications were made to allow for a difference in
equality between access and node addition. This allows for retrieval
and removal equality to be different from add equality. This should
allow for treemap and treeset to be built on top of this library.